feat: delete tournament and confirmation prompts

+++ pages/admin/judges.vue
@@ -6,7 +6,7 @@
     </div>
     <div class="m-4 py-4">
       <div class="pb-4">
-        <button class="btn btn-outline btn-sm btn-primary" @click.prevent="adding = true" :disabled="inAction">
+        <button class="btn btn-outline btn-sm btn-primary" @click.prevent="showAdd" :disabled="inAction">
           <span>Add Judge</span>
         </button>
       </div>
@@ -27,7 +27,7 @@
             <td>{{ getLevelName(t.rank) }}</td>
             <td>{{ getProvinceName(t.region) }}</td>
             <td>
-              <button class="btn btn-error btn-square btn-sm" @click.prevent="remove(t.id)" :disabled="inAction">
+              <button class="btn btn-error btn-square btn-sm" @click.prevent="showRemove(t.id)" :disabled="inAction">
                 <XMarkIcon class="w-4 h-4" />
               </button>
             </td>
@@ -35,41 +35,35 @@
         </tbody>
       </table>
     </div>
-    <div class="modal modal-bottom sm:modal-middle" :class="adding ? 'modal-open' : ''">
-      <div class="modal-box">
-        <div class="form-control w-full">
-          <label class="label" for="name">
-            <span class="label-text">Name</span>
-          </label>
-          <input id="name" name="name" type="text" class="input input-bordered" v-model="newJudge.name" />
-        </div>
-        <div class="form-control w-full">
-          <label class="label" for="province">
-            <span class="label-text">Province</span>
-          </label>
-          <select id="province" class="select select-bordered" v-model="newJudge.region">
-            <option v-for="province of Object.keys(PROVINCE_MAP)" :value="province">
-              {{ getProvinceName(province) }}
-            </option>
-          </select>
-        </div>
-        <div class="form-control w-full">
-          <label class="label" for="level">
-            <span class="label-text">Level</span>
-          </label>
-          <select id="level" class="select select-bordered" v-model="newJudge.rank">
-            <option v-for="level of Object.keys(LEVEL_MAP)" :value="level">{{ getLevelName(level) }}</option>
-          </select>
-        </div>
-        <div class="modal-action">
-          <button for="add-t-modal" class="btn btn-sm btn-error btn-outline" @click.prevent="adding = false"
-            :disabled="inAction">
-            Cancel
-          </button>
-          <button for="add-t-modal" class="btn btn-sm btn-success" @click.prevent="add" :disabled="inAction">Add</button>
-        </div>
+    <Prompt name="add_judge_modal" @submit="add" :disabled="inAction" text="Add">
+      <div class="form-control w-full">
+        <label class="label" for="name">
+          <span class="label-text">Name</span>
+        </label>
+        <input id="name" name="name" type="text" class="input input-bordered" v-model="newJudge.name" required />
       </div>
-    </div>
+      <div class="form-control w-full">
+        <label class="label" for="province">
+          <span class="label-text">Province</span>
+        </label>
+        <select id="province" class="select select-bordered" v-model="newJudge.region">
+          <option v-for="province of Object.keys(PROVINCE_MAP)" :value="province">
+            {{ getProvinceName(province) }}
+          </option>
+        </select>
+      </div>
+      <div class="form-control w-full">
+        <label class="label" for="level">
+          <span class="label-text">Level</span>
+        </label>
+        <select id="level" class="select select-bordered" v-model="newJudge.rank">
+          <option v-for="level of Object.keys(LEVEL_MAP)" :value="level">{{ getLevelName(level) }}</option>
+        </select>
+      </div>
+    </Prompt>
+    <Prompt name="delete_judge_modal" @submit="remove" text="Yes">
+      <span>Delete this judge?</span>
+    </Prompt>
   </div>
 </template>
 
@@ -83,9 +77,9 @@ const cookie = useCookie('jkj', { default: () => ({}) });
 
 const error = useState('error', () => '');
 const judges = useState('judges', () => ({}));
-const newJudge = useState('new-judge', () => (DEFAULT));
-const adding = useState('adding', () => false)
 const inAction = useState('in-action', () => false);
+const newJudge = useState('new-judge', () => (DEFAULT));
+const judgeToDelete = useState('judge-to-delete', () => undefined);
 
 const headers = { authorization: `Bearer ${cookie.value.adminCode}` };
 
@@ -99,9 +93,13 @@ try {
   inAction.value = false;
 }
 
+function showAdd() {
+  add_judge_modal.showModal();
+}
+
 async function add() {
-  const body = newJudge.value;
   try {
+    const body = newJudge.value;
     const result = await $fetch(`/api/judges`, { method: 'POST', body, headers });
     judges.value.push(result);
     newJudge.value = DEFAULT;
@@ -110,19 +108,26 @@ async function add() {
     error.value = handleServerError(err);
   } finally {
     inAction.value = false;
-    adding.value = false;
   }
 }
 
-async function remove(id) {
-  try {
-    inAction.value = true;
-    judges.value = await $fetch(`/api/judges/${id}`, { method: 'DELETE', headers });
-    error.value = '';
-  } catch (err) {
-    error.value = handleServerError(err);
-  } finally {
-    inAction.value = false;
+async function showRemove(id) {
+  judgeToDelete.value = id;
+  delete_judge_modal.showModal();
+}
+
+async function remove() {
+  if (judgeToDelete.value) {
+    const id = judgeToDelete.value;
+    try {
+      inAction.value = true;
+      judges.value = await $fetch(`/api/judges/${id}`, { method: 'DELETE', headers });
+      error.value = '';
+    } catch (err) {
+      error.value = handleServerError(err);
+    } finally {
+      inAction.value = false;
+    }
   }
 }
 </script>
